Reorganize rpi_* code. Needs further reorg. Maybe moved to ./rpi/ and given `package rpi` so it is found as gpio/rpi...
